line chart: scale aware scroll to zoom (#4663)
Previously, in the interactive_view, we used simple arithmatic to
propose new viewExtent based on scroll magnitude and velocity, but it
did not account for the scale. As such, we are now using scale to
compute the next extent based on the position and speed of the cursor
and scroll.