rust: use `TaggedRunMetadataValue` consistently (#4678)
Summary:
We currently wrap tagged metadata events in `GraphDefValue`s rather than
`TaggedRunMetadataValue`s. This doesn’t actually cause any user-facing
errors, because we still generate the correct metadata and the two kinds
of values are enriched into blob sequences in the same way (and we have
tests for the public API, so such an error would have been caught).

Perhaps this raises the question of whether the distinction between the
two types is useful at all. I think that it is: the *values* are treated
the same, but the associated functions `initial_metadata` are different.
Also, these *are* two different kinds of event values, so it seems more
faithful to represent them this way rather than some `BytesLikeValue`.

Along the way, we trim a bit of fat from this logic to make its essence
clearer. Specifically: the `SummaryValue` case needs to directly `match`
on the `hash_map::Entry` for “partially moved object” reasons, but the
other two cases are simpler and can just use [`or_insert_with`].

[`or_insert_with`]: https://doc.rust-lang.org/std/collections/hash_map/enum.Entry.html#method.or_insert_with

Test Plan:
The graphs demo includes a run called `tagged` that exercises this
codepath; it looks fine both before and after this change.

wchargin-branch: rust-trm-value-fix