timeseries: implement runs x-axis splitter
Please refer to #4969 for definition of the monotonic step and what
problem we are trying to solve.

This change adds integration of the user setting to detect non-monotonic
step changes and split runs into sub runs. While this change is only
applied to the new GPU line chart, it can easily transfer to the legacy
one.