timeseries: reflect smoothing setting in URL (#4686)
This change adds "smoothing" to query parameter for convenience of a
user. When user changes to the scalars smoothing setting to a
non-default value, it will get reflected to the URL in the query
parameter (not hash)_as per our state persistence doc.

Do note that existing Polymer code uses the hash which can be convenient
to users but kind of makes our set up more convoluted as both Polymer
and Angular app will be overwriting it.