http_client: add XSRF header to non-GET requests (#4691)
Summary:
All POST, PUT, and DELETE requests sent through `TBHttpClient` now
include a header `X-XSRF-Protected: 1`. The name and value are constant;
this is not a XSRF token. Instead, servers can determine from the
presence of this non-standard header that it was sent through JavaScript
(rather than a form) and thus is subject to the same-origin policy.

This is currently intended for Google-internal surfaces, and is
unrelated to #4685.

Test Plan:
Unit tests included. Test sync passes: <http://cl/358305731>.

wchargin-branch: http-xsrf-header