timeseries: add option for monotonic x
One of perennial problem of TensorBoard is zig-zaggy line chart. This
happens when a directory for a run contains event files with overlapping
step numbers in which case, TensorBoard will try to read them in order
then sort them by wall time. Our line chart, being true to the data we
are getting, does not have any special logic to break decreases in `x`
and renders a zig-zags which often confuses users.

This change adds a setting in timeseries to let user break such time
series into a fragment where we guarantee monotonic increases in `x`
within a line segment.