alerts: plumb followupAction into store (#4708)
The `Alerts` feature's views are responsible for handling and surfacing
alerts from actions in a "toast" UI. Modules may register a `followup`
action to explicitly show a "retry" style button that triggers a followup
action.

The retry button did not appear as expected, since the reducer was
only storing the `localizedMessage` field. This change includes the
optional `followupAction` as well.

Tested by manually modifying `metrics_module.ts`'s `alertFromAction`
registration to return a followup action, then after clicking the 'Pin'
icon on a card to trigger the action, observed that the followup action
produces the desired retry button.