graph: create initial README with user instructions (#4704)
This introduces a README in `tensorboard/plugins/graph/README.md`,
similar to other plugins. The main motivation is to:

- Catch users inspecting the repo who may want to know how the Graph
  dashboard works, and may have missed the tutorial.
- Have a place where we can add technical details that document the
  dashboard's internals.

It is **not** meant to be a replacement to the TensorFlow documentation
site's guide.  The official tutorial contains a Colab, is more stable (no TF1
references), and no implementation details of the dashboard's internals.

A followup change expands upon the technical details section:
https://github.com/tensorflow/tensorboard/pull/4705