line chart: improve major axis render
In very limited use cases of the wall time or temporal scale, the new
line chart shows a major axis to avoid overcrowding of the minor ticks
with long labels.

Previously, the major tick labels were placed only on the "beginning"
which often can be outside of the viewport. For example, for minor ticks
from "4:23" to "4:59", if the major tick goes from "<Date> 4:00" and
"<Date> 5:00", you may not be able to see the label for "4:00" at all.

This change fixes that by putting a border around, using previous
example, 4:00 to 5:00, and centering the label in the available space.