line chart: improve small number on linear scale
When you zoom in a lot on a line chart, current line chart tries to
cramp labels with many digits into the small space causing them to
overlap and cause major visual jank.

With recent improvements to the major ticks to work on y-axis too, when
we detect small differences in the extent, we now show a major ticks for
the linear scale, too, and remove a lot of redundant long prefixes from
the minor ticks.