line chart: support manual extent changes (#4711)
One of the requested feature in the scalars chart is being able to
change the axis extent manually. For example, when plotting a
probability, you want to see the chart from [0, 1] since there is
absolutely no reason to look beyond that. This change gives a UI
affordance to do so.

Clicking on an axis, you will see a popover with controls that allows
you to manually set the extents.