infra: allow SVG assets to be composed
We have different variants of TensorBoard and they _can_ have different
set of SVG icons they need. As such, currently, each product defines set
of mat-icon SVGs they require but, sadly, they don't compose.

This results in a problem because the OSS TensorBoard which is used as
the basis for the other ones can include more icons to be used but they
are not reflected accordingly on other products leading to missing
assets.

By default, Angular's svg_bundle does not allow composition but our
version does allow composition.