graph: fix pixelated SVG for Chromium (#4748)
Starting in Chromium 89, zooming in on nodes in the Graph dashboard
yielded very pixelated rendering artifacts. See details in [1]. This
regression is a result of changes to Chromium's SVG renderer [2].

To workaround the symptom, this change replaces the
`will-change: transform` CSS with `transform: translateZ(1px)`. Newer
Chromium versions take the former `will-change` as a hint to use the
rasterized result at initial scale, which renders faster at the cost of
image quality.

Note that we intentionally do not do `transform: translateZ(0)`. The
`1px` is important to avoid introducing a blurry text regression in
Firefox [4], for unknown reasons. The "#root" node does not have
`transform: translateZ(1px)` because this would interfere with
panning the graph.

The `will-change` rule was added in TB PR [3], but removing it has no
noticeable effect on load performance or panning the viewport.

[1] https://github.com/tensorflow/tensorboard/issues/4744
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=1185435
[3] https://github.com/tensorflow/tensorboard/pull/4495
[4] https://imgur.com/a/irk0jB7