data_provider: add `GetExperiment` RPC (#4750)
Summary:
This new RPC is intended to back both the `data_location` and the
`experiment_metadata` data provider methods (which really should be
combined into one). It takes an experiment ID and returns all available
metadata. If needed, we could add a fieldmask, but we leave that out
until we have a need for it.

The Rust implementation of the data server will only populate the data
location field, but we still include the rest to mirror the Python
interface.

This is the first use of `google.protobuf.Timestamp` or any of the
well-known types, so some build tweaks are needed.

Test Plan:
The protos and Rust code all build, and the newly updated instructions
in `DEVELOPMENT.md` for still work. With `grpc_cli`, the new RPC can be
seen to successfully return a non-OK `"not yet implemented"` status.

wchargin-branch: data-getexperiment-rpc