core: dump some debug info in `/data/environment`
Summary:
In #4751, we implemented `data_location` support for RustBoard, which
means that the location no longer includes the gRPC address. But while
this is an improvement for users, the address was convenient for
development, and there’s no other way to get it from the UI. This patch
augments `/data/environment` to dump some basic debug info: the
TensorBoard version, the data provider, and the command-line flags.

This behavior can be turned off. It’s enabled in `tensorboard(1)`, but
disabled by default on the `CorePlugin` and `CorePluginLoader` APIs.
Thus, users who build their own TensorBoards will not find themselves
unexpectedly leaking information about their setup. For normal users of
`tensorboard(1)`, the flags are fine to render; the only potentially
sensitive thing is the logdir path, which is of course already
prominent.

Sample response, passed through `jq .`:

```
{
  "data_location": "logs/mnist/",
  "window_title": "",
  "debug": {
    "version": "2.5.0a0",
    "data_provider": "GrpcDataProvider(addr='localhost:42127')",
    "flags": {
      "logdir": "logs/mnist/",
      "logdir_spec": "",
      "host": null,
      "bind_all": true,
      "port": null,
      "reuse_port": false,
      "load_fast": true,
      "grpc_creds_type": "local",
      "grpc_data_provider": "",
      "purge_orphaned_data": true,
      "db": "",
      "db_import": false,
      "inspect": false,
      "version_tb": false,
      "tag": "",
      "event_file": "",
      "path_prefix": "",
      "window_title": "",
      "max_reload_threads": 1,
      "reload_interval": 5,
      "reload_task": "auto",
      "reload_multifile": null,
      "reload_multifile_inactive_secs": 86400,
      "generic_data": "auto",
      "samples_per_plugin": {
        "scalars": 0,
        "images": 0
      },
      "custom_predict_fn": "",
      "wit_data_dir": "",
      "__tensorboard_subcommand": "serve"
    }
  }
}
```

Test Plan:
Manually navigated to `/data/environment` and verified that the flags
render and the data provider appears for both `--load_fast` and the
legacy multiplexer. Checked that when the change to `default.py` is
reverted, no debug data shows up.

wchargin-branch: environment-debug-info
wchargin-source: 772a3783d6f19e4060b37fe43076b963cfd8d96e