server_ingester: perform tilde expansion on logdir (#4756)
Summary:
The `rustboard` binary, as a well-behaved CLI application, does not
perform tilde expansion on its arguments, because that is the job of the
shell. However, TensorBoard historically _has_ performed tilde
expansion, and so, when `--logdir '~/foo'` is given, `--load_fast` is a
breaking change.

This patch teaches `server_ingester.py` to also perform tilde expansion.
This way, we recover the same user experience without tarnishing the
integrity of `rustboard` itself, or my precious sensibilities.

Test Plan:
Run `tensorboard --logdir '~/tensorboard_data/mnist'` or something: note
the quotes, preventing tilde expansion in the shell. Note that this
resolves to the same directory as if it had not been quoted, as visible
from the data location in the UI. Note that `--load_fast` now exhibits
the same behavior.

wchargin-branch: server-ingester-tilde-expansion