graph: notify consumers when Hierarchy's templates update (#4741)
The motivation behind these changes is to allow the `.templates` field,
currently stored in the Hierarchy, to be computed lazily, e.g. only if/when
the user selects the "Color by structure" mode.

This introduces a "Dispatcher" class, used to give the `Hierarchy` the
traits to allow it to dispatch events and let Hierarchy clients register to
listen to those events. Doing so allows components inside `<tf-node-info>`
and `<tf-graph-scene>`, who currently read the templates, to listen for
changes.