graph: color large graphs by 'none' instead of 'structure' by default (#4742)
The "Color by structure" mode can be very expensive to compute,
often consistently freezing TensorBoard for minutes with some large
graphs. This introduces a new "Color by none", which does not color
groups by structure.

Opening TensorBoard's Graphs with a small graph will still default to
"structure" coloring, but when loading a large graph, it will
automatically switch to "none", requiring the user to opt-in by clicking
on the radio button.

The condition for a large graph is "10K+ nodes && 10K+ edges", but
this can be tuned. These were determined by examining several
graphs manually.