custom_scalar: add generic data support
Summary:
This patch replaces the multiplexer code in the custom scalars plugin
with data provider code. It is not gated behind a flag. We treat the
layout protos as tensors, since they are of scalar shape and are
expected to be small. A case could be made for representing them as blob
sequences with a datacompat shape conversion, but this is simpler and
the plugin doesn’t get a ton of use.

Test Plan:
The dashboard still works with the standard demo data, both with and
without `--load_fast`.

wchargin-branch: custom-scalars-generic
wchargin-source: 91001190a543e99ba73c407df3bc801d81eaa2cf