histogram: remove unnecessary downsampling (#4776)
Summary:
As the comment says, the histograms plugin used to manually downsample
because the multiplexer data provider didn’t. As of #3272, it does, so
this code is vestigial.

Test Plan:
Note that changing the default histogram sampling constants (in the same
file) to very small numbers causes the frontend to show less data. Unit
tests for the histograms plugin also cover this: removing downsampling
from the multiplexer data provider causes the histogram tests to fail.

wchargin-branch: histogram-rm-downsample