line chart: conditionally disable fit to domain (#4758)
The new line chart now differentiate the state between viewBox overridden
vs. default. When it is put on the default, any changes in the data will
automatically be reflected in the new domain calculation while when
overriden, the viewBox will be "stuck".

This was one of the older feature requests from the users where they
wanted the line chart to clearly indicate whether it will "auto scroll"
to fit the data or it will be stuck.

With this change, we will improve the feature in a minor way by
disabling the "fit to domain" button when it is already at the default
to indicate that the data updates will trigger a domain update as well.