npmi: fix the undefined in the selector (#4779)
`getEmbeddingDataSet` is typed as something that can return an undefined
but component is typed as non-nullable which is incorrect. Angular does
not correctly type check across the component tree and resulted in
incorrect typing.

This change fixes the case when we try to access property in an
undefined in a template.