time series: change visibility to include uniqueId
TimeSeries redux store currently holds onto the visibility state of the
cards using cardId as the key. However, the problem is, there can be
multiple cards with the some `cardId` depending on the context and it
leads to suble bugs. Today, in the timeseries dashboard, the filtered
view and the main view are co-existing views that render cards with the
same cardId. However, since the DOM visibility is mutually exclusive,
i.e., while one is visible another is not visible, our store gets very
confused.

We can work around this but, still, we need to properly store these
visibility state corerctly. For instance, our scalar card defers
stamping the heavy line chart component until it is first visible on the
screen. However, if the visibility state is incorrect, there can be
caases when it has to be stamped onto the DOM while it is actually
invisible in the screen because there is another card with the same
cardId visible on the screen.

This change tries to address this by recognizing that the visibility
state is closely tied to the DOM and tries to store the state keyed by a
unique identifier specified by the view (with an ability to relate that
ID to a cardId).