time series: download link
With this change, when user clicks on overflow menu, one will find a
menu item for downloading the data. When user clicks on it, time series
will bring up a dialog which user can interact with to select a run to
download for the tag selected.

Do note that this is not yet implemented for the histograms and image
data types.