program: fall back from data server to multiplexer (#4794)
Summary:
Under `--load_fast=auto`, if the data server exits without writing a
port file, we now gracefully degrade to the multiplexer reading path.
This subsumes and expands upon the checks from #4786, with semantics now
defined by the data server as in #4793.

This depends on the `--error-file` data server flag added in #4793,
which is not yet released. To handle this, we add a mechanism for gating
flags against data server versions. As [discussed in #4689][1], this
raises a concern of what to do when the data server is chosen by a
mechanism other than the Python package. We resolve this by simply
treating such data servers as bleeding-edge. If you link in a data
server yourself, you should use it with a copy of `tb-nightly` built
from the same Git commit, or your mileage may vary.

See also #4786, which first proposed this functionality and offered some
alternative mechanisms.

[1]: https://github.com/tensorflow/tensorboard/pull/4689#discussion_r580583917

Test Plan:
Set your `GOOGLE_APPLICATION_CREDENTIALS` to an invalid JSON file, and
try running with `--load_fast=auto` with release v0.4.0 of the data
server and also with latest head (`//tensorboard/data/server:install`).
Note with `--verbosity 0` that with the old server, `--error-file` is
not passed and the data server has to fall back to anonymous creds,
whereas with the new server, `--error-file` is passed and TensorBoard
correctly falls back to the multiplexer.

wchargin-branch: data-server-fallback