ci: install TensorBoard deps after TensorFlow (#4807)
Summary:
Installing `tf-nightly` pulls in `tb-nightly` and its dependencies. This
installs over our `requirements.txt`, which matters when there are
version differences. In particular, a PR that upgrades a dependency in
`requirements.txt` should be tested on the new version, not the old one.

Test Plan:
A prior version of this PR also upgraded a package version. The “Check
Pip state” action showed that we now test on the upgraded version.

wchargin-branch: ci-fix-deps-install-order