timeseries:customizable settings w/ DI
This change introduces DI for the time series' redux state to allow
different application to provide the default settings. Do note that this
pattern CANNOT be used for a route aware states.

Also do note that this change introduces some infrastructure for
integrational style testing using Jasmine to better test the behavior
provided by the metrics_module which otherwise is not really tested.