cli: enable fast Rust data loading by default (#4813)
Summary:
Running `tensorboard` now defaults to the fast data loading codepath
when available. This *does not* affect systems where RustBoard is not
installed, including on platforms where it’s unavailable. It also does
not affect invocations in which RustBoard detects that it is not
suitable for the job, such as an unsupported remote filesystem or
credential set. In such cases, TensorBoard will automatically fall back
to the legacy multiplexer paths. This selection is silent unless
`--verbosity 0` is given.

Because of this automatic fallback detection, this change is expected to
be “safe”. There may of course be undiscovered bugs in RustBoard, in
which case users have `--load_fast=false` as an escape hatch.

Users can also set `--load_fast=true` to bypass the selection logic,
suppress the advisory note that new loading logic is being used, and
hard-fail if RustBoard is not available.

Test Plan:
`bazel run //tensorboard` loads fast.

wchargin-branch: load-fast-auto