histograms: warn when metadata version is too new
Summary:
We’re considering making changes to the histogram data format, most
immediately to resolve dynamic shape issues, but maybe also to generally
improve the sampling and binning behavior. This patch teaches the
histograms plugin to check the summary metadata’s `version` field and
complain if it’s not not `0` (the default, and the implicit value).
This way, if users try to use an old TensorBoard version to read data in
the new format, they’ll see a nice message instead of potential crashes
or misrepresented data.

It’d be reasonable to adopt something similar for all our plugins, but
let’s start here for now.

Test Plan:
Changing the check to `md.version != 777` properly causes a message to
be printed once (only) and no data to appear in the UI.

wchargin-branch: histograms-version-warning
wchargin-source: efe5f3e292e6db0c06cd2782c007cff9341cee41