infra: upgrade FE dependencies
This simply upgrades all the transitive dependencies using `yarn
upgrade-interactive` using the version spec specified in the
package.json.

Do note that this change upgrade `y18n@4.0.0` (to 5.0.5) which is known
to have a security vulnerability, CVE-2020-7774.

Also do note that we cannot upgrade everything including d3 as their
types have changed and is somewhat hard to debug what the underyling
issue is.