rust: hold commit in `Arc` instead of leaking (#4834)
Summary:
While watching @jonhoo’s latest [“Crust of Rust” video on atomics][jh],
I realized that we can just convert this irritating `&'static Commit`
leaked reference into a jointly owned `Arc<Commit>` while still meeting
the `T: 'static` bound for the RPC handler type. It’s not really a big
functional difference, because the commit would live for the whole
program anyway, but it’s nice to not have to worry about it. It should
also reduce the peak memory usage during tests, since we used to box up
and leak a new commit in each test case.

(In fact, this probably makes server shutdown a tad slower. If we really
wanted to recover that performance, we could make the server generic
over `T: Deref<Target = Commit>`, and pass `Box::leak(...)` in the main
program and an `Arc<Commit>` in tests. But that seems overkill; dropping
a few hash maps is probably not the end of the world.)

[jh]: https://youtu.be/rMGWeSjctlY?t=1181

wchargin-branch: rust-no-leak-commit