pip: loosen `pandas` dep to 1.x
Summary:
Running `pip install pandas==1.0.3` on Python 3.9 hangs for a few
minutes and then crashes with 20K lines of C compilation errors. The
latest version of `pandas` works with 3.9, but not with 3.6, which we
also support (and run in CI). This patch loosens our dep to `~=1.0`,
which should pick an appropriate version in both 3.6 and 3.9.

wchargin-branch: pip-pandas-1.x
wchargin-source: 4e2a4a69f57edf9f08bdc3ddfc6c722c79b0a5fe