plugin api: refactor to use Angularism (#4839)
Currently the PluginApis are implemented a bit funnily; it uses
Polymer's RunsStore but exposes API to Angular via a Polymer component
hack.

This is not ideal for our long term code health, testability, and API
expandability. This change refactors to use NgModule, DI, and Angular
view patterns to better implement the host side of the APIs.