notification center: integrate markdown renderer
This change integrates the markdown renderer with the notification
center. Now the content of the notification can be any markdown string
as long as it does not contain certain dangerous elements like script
tag.

This change also makes few visual modification to the markdown renderer
for the optimal rendering in the small widthed screen.