time series: move the fit button to header (#4856)
The scalar line chart in the time series has had the feature where we
automatically fit domain to the new data when it is unaltered. However,
because it was hidden inside the overflow menu and because there was no
clear visual indication when the chart will fit to the domain or not, it
was largely confusing. With the recent changes to make it more salient
with `title` and `disabled` on the button, we improved in this area but
was still not too obvious.

To fix this problem, we are now moving the control out of the overflow
menu into the chart header so there is a clearer visual difference
between domains overridden vs. not.