infra: update JS dependencies.
This change upgrade:
- Angular
- Rxjs
- rules_nodejs
- rules_sass
- @bazel/*
- @bazel/karma -> @bazel/concatjs

This change does not include upgrade to @ngrx/* as their 10+ broke the
build as their es module-d file gets included as part of the named
modules and gets included in our requirejs builds (:dev and :karma
targets).

Do note that with the toolchain upgrade, we now have to specify the
return types of `JSON.parse` more strictly causing larger changes than
intended.