timeseries: correctly fix fit-to-domain (#4900)
In #4890, we incorrectly fixed the issue where scalar-card-component
does not update its DOM based on changes.

Background: while gory details are not yet known personally to me,
Angular, roughly speaking, has very strict condition in which it will
update the DOM. To name a few I know, they are (1) browser event, (2)
async (e.g., setTimeout), and (3) async pipe. In the earlier change, we
assumed that the `@Output` is one of them but it does not appear so.

Assumption: Angular updates DOM top to bottom (component tree wise) and
even if child component communicate information upwards when it mounts,
it _can_ be too late for parent component to actually apply the update
to the DOM.

Fix: we now use the AsyncPipe and Observable to communicate the
information upwards from a child component.

Why is this correct?
--------------------
AsyncPipe, on next value, interacts with the change detector[1]. Also,
this is a pattern that, sometimes, CDK components use[2]. Please see
public methods like `attachments` or `backdropClick`. At first, I was
curious why CDK did not use `Output` but now it starts to click a
little.

Test Case
---------
1. Launch TensorBoard and go to TimeSeries
2. Set the tag filter so only one scalar card matches the filter
3. Type a typo so nothing matches
4. Remove that single character typo
5. Repeat 3-4 to see the fit-to-domain turn solid (without this patch)

[1]: https://github.com/angular/angular/blob/master/packages/common/src/pipes/async_pipe.ts#L146
[2]: https://material.angular.io/cdk/overlay/api