core: fetch runs when alias changes
Polymer based TensorBoard is unaware of notion of an experiment and to
naively support it, we have made changes to the runs endpoint so it
returns a unique name of a run per experiment by prefixing name of an
experiment. However, it is not as simple as just that.

Because name of an experiment can be arbitrary large and can contain
illegal character like "/", we have a notion of alias per experiment.

The run-selector is unaware of anything experiment simply knows about
list of run names (not ids) and communicates with other Polymer based
components using the run names. As a result, whenver alias changes, we
need to refresh the runs data and reload the whole Polymer dashboard so
all components display and speak the same run name.

This change triggers Polymer run load whenever alias changes.