refactor: use router event to fetch data (again) (#4941)
Revert "Revert "refactor: use router event to fetch data. (#4915)""

This reverts commit 32e533f8d965036867233a59ffb25bb6bdf2fb97.

Original PR message
================
Context: previously, we were using a composite action[1] to know when the
dashboard view updated and thus its data needs to be refreshed.

Fix: now that TensorBoard has a router and can determine what views are
rendered given a route, we can use a purer redux pattern to determine
whether the dashboard data has to be reloaded. Since not all versions of
the app uses the router, we will temporarily support the coreLoaded.

[1]: composite actioness: view was subscribing to the state changes to
determine to fire an action which in turn triggers another action.