runs: refactor to move color state to data
Intention: when the color grouping changes, we need to reassign the
colors for every run and, to do that, we need current list of runs. In
the `ui` state, we merely have settings without any informations about
which runs we have for a given experiment.

This change refactors the color state to the `data` reducers so we can
refer to the list of runs for this experiment.

Alternative: change the shape of the `data` and colors so it is keyed by
route id and runIds without using the routeContextedUtil. While that
would keeps state structure a bit more static, it makes downstream
selectors all specify `experimentIds`, making it harder to use and
refactor larger.