runs: implement Runs and Experiments grouping
This change implements utility for grouping by experiments and runs.
From the experiment-less users, this change should be a noop.

This change also reorganizes the default color assignment to operate
based on the group generated by the utility.

Alternative considered: thought about only remembering `groupId` to
color instead of `groupToColor` and `defaultRunColor`. With this mode,
we would read the `runIds` state and ultimately create run-to-color map
but this requires:
1. stitching information in the selector making the memoization trickier
2. users of the selector pass current `experimentIds` so we can obtain
  proper `runIds`. This would increase blast radius of the change a bit
  too large.