runs: conditionally add group menu button
This change adds the menu button for grouping runs. The button is
rendered conditionally when the feature is enabled.

Do note that this change alters the color fobs in the runs table column
as we need to horizontally align them with the new menu button
(otherwise, it looks really wonky; all in all, it is more correct to
center align them anyways).

This change yet does not integrate with actions or the enums.