Explicitly list ngrx dependencies in PluginApiHostModule. (#4879)
TensorBoard frontend team is experimenting with a new policy to explicitly import ngrx-based dependencies in NgModules that otherwise don't need to be listed.

We first apply this experimental policy to PluginApiHostModule. Recent changes to that module to depend on selectors in other parts of the codebase, and the surprising runtime breakages it created in Google-internal TensorBoard services, are what provoked this discussion in the first place.