settings: add redux state for global state
Motivation
----------
In the multi-experiment version of TensorBoard, settings in the metrics
dashboard "resets" to its default values every time a new experiment
opens. Users got a bit uncomfortable doing that so we want to improve
the UX by remembering user's inter-experiment global settings.

This change
-----------
This change simply adds the redux state but we eventually want to
remember the settings via localStorage. While this solution is not a
panacea (we still forget when user rotates port number), this can
improve certain experience drastically.

Again, this change only adds redux state, reducers, actions, and
selectors.

Discussion
----------
For consumer of the selector, it is quite hard to now mix the
globalSetting + per-experiment override. In fact, you really need three
selectors that purportedly do the similar things:
- one for getting the global setting (needed for rendering right nav).
- one for getting and rendering per-experiment setting (needed for
  settings view component).
- one for mixing the two and just getting the, for example, smoothing.

For providing the third kind of selector, we may wanto to provide a
composite selector that grabs state from two different features.
However, placement of said selector is murky and such selector creates
dependencies on multiple DataNgModule.